Moved imports, renamed, and validated training. Ready for merge IMO
